app/requirements.txt
Flask==2.2.3
torch==2.1.0
Pillow==9.3.0
numpy==1.23.5
